refactor: move login screen code to Class60
