+++ src/main/java/org/runejs/client/Class26.java
@@ -5,7 +5,6 @@ import org.runejs.client.cache.CacheArchive;
 import java.awt.*;
 
 public class Class26 {
-    public static int loginScreenState = 0;
     public static Image loadingBoxImage;
     public static CacheArchive aCacheArchive_632;
     public static int baseY;
